Add user manager

+++ smarket/src/main/java/dev/finhacker/smarket/domain/enterprise/EnterpriseRepository.java
@@ -2,6 +2,6 @@ package dev.finhacker.smarket.domain.enterprise;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EnterpriseRepository extends JpaRepository<Enterprise, Long> {
+public interface EnterpriseRepository extends JpaRepository<Enterprise, Integer> {
 
 }
